# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/poi/route": {
      "post": {
        "operationId": "PoiController_route",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutePoiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Poi"
        ]
      }
    },
    "/api/poi/reports": {
      "post": {
        "operationId": "PoiController_report",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoiReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Poi"
        ]
      }
    },
    "/api/admin/poi": {
      "post": {
        "operationId": "AdminPoiController_create[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      },
      "get": {
        "operationId": "AdminPoiController_list[0]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/v1/admin/poi": {
      "post": {
        "operationId": "AdminPoiController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      },
      "get": {
        "operationId": "AdminPoiController_list[1]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/admin/poi/{id}": {
      "patch": {
        "operationId": "AdminPoiController_update[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePoiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      },
      "delete": {
        "operationId": "AdminPoiController_remove[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/v1/admin/poi/{id}": {
      "patch": {
        "operationId": "AdminPoiController_update[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePoiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      },
      "delete": {
        "operationId": "AdminPoiController_remove[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/admin/poi/import": {
      "post": {
        "operationId": "AdminPoiController_import[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPoiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/v1/admin/poi/import": {
      "post": {
        "operationId": "AdminPoiController_import[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPoiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/admin/poi/reports": {
      "get": {
        "operationId": "AdminPoiController_listReports[0]",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/v1/admin/poi/reports": {
      "get": {
        "operationId": "AdminPoiController_listReports[1]",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/admin/poi/reports/{id}/approve": {
      "post": {
        "operationId": "AdminPoiController_approve[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeratePoiReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/v1/admin/poi/reports/{id}/approve": {
      "post": {
        "operationId": "AdminPoiController_approve[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeratePoiReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/admin/poi/reports/{id}/reject": {
      "post": {
        "operationId": "AdminPoiController_reject[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeratePoiReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
    "/api/v1/admin/poi/reports/{id}/reject": {
      "post": {
        "operationId": "AdminPoiController_reject[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeratePoiReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPoi"
        ]
      }
    },
...
  },
  "info": {
    "title": "InterCity API",
    "description": "Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RoutePoiDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePoiReportDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "speed_camera",
              "hazard",
              "police",
              "school_zone",
              "toll",
              "fuel",
              "other"
            ]
          },
          "lat": {
            "type": "number"
          },
          "lon": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "radiusMeters": {
            "type": "number",
            "minimum": 100,
            "maximum": 3000,
            "default": 1200
          }
        },
        "required": [
          "type",
          "lat",
          "lon"
        ]
      },
      "CreatePoiDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePoiDto": {
        "type": "object",
        "properties": {}
      },
      "ImportPoiItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "speed_camera",
              "hazard",
              "police",
              "school_zone",
              "toll",
              "fuel",
              "other"
            ]
          },
          "lat": {
            "type": "number"
          },
          "lon": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "radiusMeters": {
            "type": "number",
            "default": 1200
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "type",
          "lat",
          "lon"
        ]
      },
      "ImportPoiDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportPoiItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "ModeratePoiReportDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          }
        }
      },
...
    }
  }
}